# Conflicts:
#	frontend/src/containers/Destinations/DestinationCard/Carousel/Carousel.vue

+++ backend/app/controllers/DestinationController.java
@@ -18,7 +18,6 @@ import repository.DestinationRepository;
 import javax.inject.Inject;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.List;
 import java.util.concurrent.CompletionException;
@@ -52,8 +51,10 @@ public class DestinationController extends Controller {
      * @param request Http.Request the http request
      * @return a completion stage and a status code 200 if the request is successful, otherwise returns 500.
      */
+    @With(LoggedIn.class)
     public CompletionStage<Result> getDestinations(Http.Request request) {
-        return destinationRepository.getDestinations()
+        User user = request.attrs().get(ActionState.USER);
+        return destinationRepository.getDestinationsbyUserId(user.getUserId())
                 .thenApplyAsync(destinations -> ok(Json.toJson(destinations)), httpExecutionContext.current());
     }
 
@@ -65,15 +66,21 @@ public class DestinationController extends Controller {
      * @return a completion stage and a Status code of 200 and destination details as a Json object if successful,
      * otherwise returns status code 404 if the destination can't be found in the db.
      */
-
+    @With(LoggedIn.class)
     public CompletionStage<Result> getDestination(int destinationId, Http.Request request) {
-
+        User user = request.attrs().get(ActionState.USER);
         return destinationRepository.getDestinationById(destinationId)
                 .thenApplyAsync((destination) -> {
                     if (!destination.isPresent()) {
                         ObjectNode message = Json.newObject();
                         message.put("message", "No destination exists with the specified ID");
                         return notFound(message);
+                    }
+
+                    if (destination.get().getDestinationOwner() != user.getUserId()) {
+                        ObjectNode message = Json.newObject();
+                        message.put("message", "You are not authorised to get this destination");
+                        return forbidden(message);
                     }
 
                     JsonNode destAsJson = Json.toJson(destination);
@@ -245,11 +252,14 @@ public class DestinationController extends Controller {
                     }
 
                     for (int destId : duplicatedDestinationIds) {
-                        Optional<Destination> optDest = Destination.find.query().
-                                where().eq("destination_id", destId).findOneOrEmpty();
-                        destinationRepository.deleteDestination(optDest.get().getDestinationId());
+                        List<DestinationPhoto> photoDest = DestinationPhoto.find.query()
+                                .where().eq("destination_destination_id", destId).findList();
+                        destinationRepository.deleteDestination(destId);
 
-                        // TODO: Transfer the destinationId to the public destinationId
+                        for (DestinationPhoto photo : photoDest) {
+                            photo.setDestination(destination);
+                            destinationRepository.insertDestinationPhoto(photo);
+                        }
                     }
 
                     return destinationRepository.update(destination);
@@ -283,12 +293,18 @@ public class DestinationController extends Controller {
      */
     @With(LoggedIn.class)
     public CompletionStage<Result> deleteDestination(int destinationId, Http.Request request) {
+        User user = request.attrs().get(ActionState.USER);
         return destinationRepository.getDestinationById(destinationId)
                 .thenComposeAsync(optionalDestination -> {
                     if (!optionalDestination.isPresent()) {
                         throw new CompletionException(new NotFoundException());
                     }
                     Destination destination = optionalDestination.get();
+                    System.out.println(destination.getDestinationOwner());
+                    System.out.println(user.toString());
+                    if (destination.getDestinationOwner() != null && destination.getDestinationOwner() != user.getUserId()) {
+                        throw new CompletionException(new ForbiddenRequestException("You are not permitted to delete this destination"));
+                    }
                     ObjectNode success = Json.newObject();
                     success.put("message", "Successfully deleted the given destination id");
                     return this.destinationRepository.deleteDestination(destination.getDestinationId());
@@ -301,7 +317,12 @@ public class DestinationController extends Controller {
                         ObjectNode message = Json.newObject();
                         message.put("message", "The given destination id is not found");
                         return notFound(message);
+                    } catch (ForbiddenRequestException forbiddenReqE) {
+                        ObjectNode message = Json.newObject();
+                        message.put("message", forbiddenReqE.getMessage());
+                        return forbidden(message);
                     } catch (Throwable serverError) {
+                        serverError.printStackTrace();
                         return internalServerError();
                     }
                 });
@@ -385,7 +406,6 @@ public class DestinationController extends Controller {
                                     throw new CompletionException(new NotFoundException("Could not find photo"));
                                 }
 
-                                //TODO test this
 
                                 if (    optionalDestination.get().getIsPublic() &&                          // The destination is public
                                         optionalDestination.get().getDestinationOwner() != null &&          // The owner is not already null
